requirements.txt
async-irc
pyyaml
influxdb
